Hecho con JuampiRombola. Agregados los metodos equals y calcularDistancia a la clase Posicion con sus tests y modificada el tipo de rango en Arma (se realizo un casteo a Double)

+++ src/unidades/Arma.java
@@ -10,8 +10,8 @@ public class Arma {
 		this.danio = danio;
 		this.rango = rango;
 	}
-	public void atacar(Atacable atacable, int distanciaEntreArmaYAtacable){
-		if(distanciaEntreArmaYAtacable <= rango){
+	public void atacar(Atacable atacable, double distanciaEntreArmaYAtacable){
+		if(distanciaEntreArmaYAtacable <= (double)rango){
 			atacable.recibePuntosDeDanio(danio);
 		}
 	}
